refactor: replace all engine implementations with optimized Smol
- Remove all engine variants (Rayon, Tokio, basic Smol)
- Replace main SearchEngine with OptimizedSmol implementation
- Remove engine selection logic from main.rs
- Delete benchmark binaries for removed engines
- Simplify codebase to single optimized implementation
- Enable smol feature by default in Cargo.toml
