--- src/bin/bench_rayon_v3.rs
@@ -1,15 +1,0 @@
-use anyhow::Result;
-use ccms::{parse_query, SearchOptions};
-use ccms::search::optimized_rayon_engine_v3::OptimizedRayonEngineV3;
-
-fn main() -> Result<()> {
-    let pattern = "~/.claude/projects/**/*.jsonl";
-    let query = parse_query("claude")?;
-    let options = SearchOptions::default();
-    
-    let engine = OptimizedRayonEngineV3::new(options);
-    let (_results, _duration, total) = engine.search(pattern, query)?;
-    println!("Found {} results", total);
-    
-    Ok(())
-}
