+++ src/lib.rs
@@ -11,6 +11,3 @@ pub use schemas::{SessionMessage, ToolResult};
 pub use search::{
     SearchEngine, default_claude_pattern, discover_claude_files, expand_tilde, format_search_result,
 };
-
-#[cfg(feature = "async")]
-pub use search::{AsyncSearchEngine, AsyncSearchOptions, AsyncSearchResult};
